add pages and rearrange App
