+++ src/pages/About.jsx
@@ -1,0 +1,12 @@
+import "./../App.css";
+import React from "react";
+
+function About() {
+  return (
+    <div className="About">
+      <h1>About</h1>
+    </div>
+  );
+}
+
+export default About;
